+++ llm_mcp_server/controllers/__init__.py
@@ -1,1 +1,2 @@
 from . import mcp_controller
+from . import mcp_oauth_controller